requirements.txt
azure-common==1.1.23
azure-eventhub==5.0.0b2
certifi==2019.6.16
PyYAML==5.1.2
six==1.12.0
uamqp==1.2.2
pylint==2.3.1
